+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.12.0
+
+- Add `analysis_options` and check for unnecessary `new` and `const` keywords
+- Remove dependency on `serializable_core`
+- Add dependency to `built_mirrors`
+- Change builder from `PartBuilder` to `SharedPartBuilder`, this way generator can run alongside with other generators
+- Make builder `auto_apply: dependents`. This way builder only run on direct dependents
+
 ## 0.11.3
 
 - upgrade `analyzer` to version `^0.36.0`